Added static cast for 0 in std::accumulate call. Formatting changes.

+++ code_samples/cpp/thread_sum/sum_ints.cc
@@ -9,7 +9,7 @@
 namespace holistic_multithreading {
 namespace {
 
-void SumValues(int64_t from, int64_t to, int64_t& result) {
+void SumValues(int64_t from, int64_t to, int64_t &result) {
   for (int64_t i = from; i <= to; i++) {
     result += i;
   }
@@ -38,13 +38,15 @@ int64_t MultithreadedSum(int64_t from, int64_t to, int64_t thread_count) {
   for (int64_t i = 0; i < thread_count; i++) {
     int64_t end = (i < thread_count - 1) ? start + batch_size - 1 : to;
     results.push_back(0);
-    threads.push_back(std::thread(SumValues, start, end, std::ref(results.back())));
+    threads.push_back(
+        std::thread(SumValues, start, end, std::ref(results.back())));
     start = end + 1;
   }
-  for (auto& t : threads) {
+  for (auto &t : threads) {
     t.join();
   }
-  return std::accumulate(results.begin(), results.end(), 0);
+  return std::accumulate(results.begin(), results.end(),
+                         static_cast<int64_t>(0));
 }
 
 int64_t SingleThreadSum(int64_t from, int64_t to) {
@@ -53,6 +55,6 @@ int64_t SingleThreadSum(int64_t from, int64_t to) {
     result += i;
   }
   return result;
-}  
+}
 
 } // namespace holistic_multithreading
